Initial HTML file, JavaScript file and css file
This initial commit contains the index.html file where the calculator UI is. A form consisting of 3 inputs with their respective labels, and a button that triggers the form to submit.
The JavaScript file contains:
a shell/empty function called 'calculate' always returning 0
a function to handle form submission, reading in the input values and calling the calculate function with those values.